Back out "Callbox: allow addition or deletion of authorizers at runtime."
Original commit changeset: 361410559c07
