+++ MODULE.bazel
@@ -148,7 +148,6 @@ use_repo(
     "com_github_aws_aws_lambda_go",
     "com_github_aws_aws_sdk_go_v2",
     "com_github_aws_aws_sdk_go_v2_config",
-    "com_github_aws_aws_sdk_go_v2_service_dynamodb",
     "com_github_aws_aws_sdk_go_v2_service_s3",
     "com_github_awslabs_aws_lambda_go_api_proxy",
     "com_github_bazelbuild_bazel_watcher",
