+++ ts/pulumi/zemn.me/api/spec.yaml
@@ -15,77 +15,16 @@ components:
       type: openIdConnect
       openIdConnectUrl: "https://accounts.google.com/.well-known/openid-configuration"
   schemas:
-    Error:
-      type: object
-      required:
-        - cause
-      properties:
-        cause:
-          type: string
-          example: Too many concurrent requests.
     GetPhoneNumberResponse:
       type: object
       required:
-        - phoneNumber
+          - phoneNumber
       properties:
         phoneNumber:
           type: string
           example: "+447823434620"
-    PhoneNumberList:
-      type: array
-      items:
-        type: string
-    PhoneNumberPatchRequest:
-      type: object
-      properties:
-        add:
-          $ref: '#/components/schemas/PhoneNumberList'
-        remove:
-          $ref: '#/components/schemas/PhoneNumberList'
 
 paths:
-  /callbox/authorizers:
-    patch:
-      description: >
-        Add or remove people from the list of authorizers for entry,
-        by phone number.
-      security:
-        - googleOIDC: []
-      requestBody:
-        description: List of phone numbers to add or remove
-        required: true
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/PhoneNumberPatchRequest'
-      responses:
-        "200":
-          description: >
-            Returns an updated list of phone numbers constituting
-            the full list of approvers.
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/PhoneNumberList'
-    get:
-      security:
-        - googleOIDC: []
-      responses:
-        "500":
-          description: >
-            Something went wrong.
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/Error'
-        "200":
-          description: >
-            Returns a list of phone numbers representing users
-            who are authorised to approve entry.
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/PhoneNumberList'
   /phone/number:
     get:
       summary: Returns the current phone number for the callbox.
